+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue #18 : Improved error reporting by returning the ErrorReport
 ### Fixed
 - Issue #20 : Fixed index-out-of-range panic while getting vaultRecord  
+- Issue #21 : Can't set expire / end date for vault records
 
 ## [1.2.3] - 2022-11-07
 ### Fixed